Random Fixes

+++ Aphro-WebForms/Models/SearchResponse.cs
@@ -2,7 +2,7 @@
 {
     public class SearchResponse
     {
-        public long xid { get; set; }
+        public int xid { get; set; }
         public string firstname { get; set; }
         public string lastname { get; set; }
     }
